+++ src/versions.ts
@@ -6,22 +6,27 @@
 
 import * as childProcess from 'child_process';
 import * as crypto from 'crypto';
-import * as fsExtra from 'fs-extra';
+import fsExtra from 'fs-extra';
 import * as path from 'path';
 import * as util from 'util';
+import {createRequire} from 'module';
+const require = createRequire(import.meta.url);
 
 const execFilePromise = util.promisify(childProcess.execFile);
 const execPromise = util.promisify(childProcess.exec);
 
-import {MountPoint} from './server';
+import {MountPoint} from './server.js';
 import {
   BenchmarkSpec,
   GitDependency,
   NpmPackageJson,
   PackageDependencyMap,
   PackageVersion,
-} from './types';
-import {fileKind, runNpm, throwUnreachable} from './util';
+} from './types.js';
+import {fileKind, runNpm, throwUnreachable} from './util.js';
+
+import * as url from 'url';
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 interface GitDependencyWithTempDir extends GitDependency {
   tempDir: string;
@@ -299,7 +304,8 @@ const installSuccessFile = '__TACHOMETER_INSTALL_SUCCESS__';
  */
 export async function prepareVersionDirectory(
   {installDir, packageJson}: NpmInstall,
-  forceCleanInstall: boolean
+  forceCleanInstall: boolean,
+  npmrc?: string
 ): Promise<void> {
   if (forceCleanInstall) {
     await fsExtra.remove(installDir);
@@ -319,6 +325,13 @@ export async function prepareVersionDirectory(
     path.join(installDir, 'package.json'),
     JSON.stringify(packageJson, null, 2)
   );
+  if (npmrc) {
+    await fsExtra.copy(
+      path.resolve(npmrc),
+      path.join(installDir, '.npmrc'),
+      {}
+    );
+  }
   await runNpm(['install'], {cwd: installDir});
   await fsExtra.writeFile(path.join(installDir, installSuccessFile), '');
 }