+++ CHANGELOG.md
@@ -7,6 +7,48 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- ## Unreleased -->
 
+## [0.7.1] 2024-07-18
+
+- Update jsonwebtoken dependency.
+
+## [0.7.0] 2022-07-15
+
+- [**BREAKING**] Tachometer is now an ES module binary, so it requires Node 12
+  or higher.
+- Upgraded dependencies and removed unnecessary ones.
+
+## [0.6.0] 2022-07-01
+
+- [**BREAKING**] Benchmark and `root` paths are now interpreted relative to the
+  location of the config file, instead of the current working directory.
+
+  For example, a config at path `<repo>/benchmarks/foo/tachometer.json` that
+  used to look as follows, and had to be run with `<repo>` as the cwd:
+
+  ```json
+  {
+    "root": ".",
+    "benchmarks": [
+      {
+        "url": "benchmarks/foo/index.html"
+      }
+    ]
+  }
+  ```
+
+  Should now look like this, and can be run from any cwd:
+
+  ```json
+  {
+    "root": "../..",
+    "benchmarks": [
+      {
+        "url": "foo.html"
+      }
+    ]
+  }
+  ```
+
 ## [0.5.10] 2021-09-27
 
 - The `--horizons` flag and `horizon` config setting has been replaced with